requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
DateTime==4.3
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-filter==2.4.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
humanize==3.9.0
psycopg2==2.8.6
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
zope.interface==5.4.0
